+++ WebService.API/VirtualBase/IProduct.cs
@@ -2,8 +2,8 @@ namespace WebService.API.VirtualBase;
 
 public interface IProduct
 {
-	public long Id { get; set; }
+	public int Id { get; set; }
 	public string? Name { get; set; }
 	public string? Description { get; set; }
-	public float BasePrice { get; set; }
+	public double BasePrice { get; set; }
 }
